8657 - added unit tests for generateSupportLink (#9180)
* added unit tests for generateSupportLink

* added one more test case for generateSupportLink

* use it.each method to combine separate tests that are related

---------
